Restrict sessions to same-origin requests
Our API is intentionally open to cross-origin requests, but we want to ensure that an attacker can't create a website that surreptitiously uses a logged-in user's session to make changes to the DB on that user's behalf. With this, requests from other origins (using the Referer header, which in-browser JS cannot alter) don't have access to the session, meaning they must include authentication credentials instead of piggybacking on an existing session.